+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version 2024.3
 
 * 🌟 Trakt ratings: support removing a rating.
 * 🔨 Trakt comments: refresh will actually show the latest comments, including after posting one.
+* 📝 Latest user interface translations from Crowdin.
 
 Version 2024.2
 --------------